+++ tests/test_errors.py
@@ -57,3 +57,23 @@ def test_input_date_system_error():
 def test_event_date_system_error():
     with pytest.raises(RokhValidationError, match=r"`event_date_system` must be None or an instance of DateSystem"):
         _ = get_events(year=1404, month=1, day=1, input_date_system=DateSystem.JALALI, event_date_system="Jalali")
+
+
+def test_date_error1():
+    with pytest.raises(RokhValidationError, match=r"The input date is not valid"):
+        _ = get_events(year=1404, month=12, day=31, input_date_system=DateSystem.JALALI)
+
+
+def test_date_error2():
+    with pytest.raises(RokhValidationError, match=r"The input date is not valid"):
+        _ = get_events(month=12, day=31, input_date_system=DateSystem.JALALI)
+
+
+def test_date_error3():
+    with pytest.raises(RokhValidationError, match=r"The input date is not valid"):
+        _ = get_events(year=2026, month=2, day=31, input_date_system=DateSystem.GREGORIAN)
+
+
+def test_date_error4():
+    with pytest.raises(RokhValidationError, match=r"The input date is not valid"):
+        _ = get_events(year=1445, month=12, day=31, input_date_system=DateSystem.HIJRI)